Moving `BearerRefresh` to its own component.
